WIP: creating product and sum types, extracting files from type mod
